WIP: Update to SCIFIO 0.38.x API
This commit should not need to exist. All of the breakages "fixed" here
should be addressed in the SCIFIO code, by readding methods with 'String
location' signatures, which lean on the LocationService#resolve method.
This will avoid a lot of pain downstream! And there is nothing wrong
with keeping such convenience methods around indefinitely!